Refactor event file structure
