+++ app/src/main/AndroidManifest.xml
@@ -12,11 +12,11 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".Event.EventsActivity"></activity>
+        <activity android:name=".Event.AllEvents.EventsActivity"></activity>
         <activity android:name=".ForgetPassword.ForgetPasswordActivity"></activity>
-        <activity android:name=".Event.SavedEventsActivity"></activity>
+        <activity android:name=".Event.SavedEvents.SavedEventsActivity"></activity>
         <activity android:name=".MainActivity"></activity>
-        <activity android:name=".Event.SingleEventActivity"></activity>
+        <activity android:name=".Event.SingleEvent.SingleEventActivity"></activity>
         <activity android:name=".Login.LoginActivity" android:windowSoftInputMode="adjustPan"></activity>
 
         <activity android:name=".Splash.SplashActivity">
