+++ app/src/main/res/layout/activity_all_events.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".Event.EventsActivity"
+    tools:context=".Event.AllEvents.EventsActivity"
     android:background="@color/white">
 
     <LinearLayout
