+++ app/src/main/res/layout/activity_saved_events.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:orientation="vertical"
     android:background="@color/white"
-    tools:context=".Event.SavedEventsActivity">
+    tools:context=".Event.SavedEvents.SavedEventsActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
